[Refactor] 패키지 구조 변경, 테스트 커버리지 추가
Ref : #24
