+++ src/main/java/com/fcc/PureSync/context/board/controller/BoardController.java
@@ -1,14 +1,12 @@
-package com.fcc.PureSync.controller;
+package com.fcc.PureSync.context.board.controller;
 
-import com.fcc.PureSync.dto.BoardDto;
+import com.fcc.PureSync.context.board.dto.BoardDto;
 import com.fcc.PureSync.dto.ResultDto;
 import com.fcc.PureSync.jwt.CustomUserDetails;
 import com.fcc.PureSync.repository.MemberRepository;
-import com.fcc.PureSync.service.BoardService;
-import com.fcc.PureSync.util.FileUploadUtil;
+import com.fcc.PureSync.context.board.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
