+++ src/main/java/com/fcc/PureSync/context/board/dto/BoardFileDto.java
@@ -1,7 +1,6 @@
-package com.fcc.PureSync.dto;
+package com.fcc.PureSync.context.board.dto;
 
-import com.fcc.PureSync.entity.Board;
-import com.fcc.PureSync.entity.BoardFile;
+import com.fcc.PureSync.context.board.entity.BoardFile;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
