+++ src/main/java/com/fcc/PureSync/context/board/entity/Board.java
@@ -1,6 +1,6 @@
-package com.fcc.PureSync.entity;
+package com.fcc.PureSync.context.board.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fcc.PureSync.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
