+++ src/main/java/com/fcc/PureSync/context/board/entity/Likes.java
@@ -1,40 +1,30 @@
-package com.fcc.PureSync.entity;
+package com.fcc.PureSync.context.board.entity;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fcc.PureSync.context.board.entity.Board;
+import com.fcc.PureSync.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 @Entity
-@Table(name = "tb_comment")
+@Table(name = "tb_likes")
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class Comment {
+public class Likes {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long cmtSeq;
-    private String cmtContents;
-    @Builder.Default
-    private LocalDateTime cmtWdate=LocalDateTime.now();
-    @Builder.Default
-    private Integer cmtStatus=1;
-
+    private Long likesSeq;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "mem_seq")
     private Member member;
 
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "board_seq")
     private Board board;
-
-
-
 }
