+++ src/main/java/com/fcc/PureSync/context/board/repository/BoardRepository.java
@@ -1,15 +1,13 @@
-package com.fcc.PureSync.repository;
+package com.fcc.PureSync.context.board.repository;
 
-import com.fcc.PureSync.entity.Board;
+import com.fcc.PureSync.context.board.entity.Board;
 import com.fcc.PureSync.entity.Member;
-import com.fcc.PureSync.entity.Likes;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface BoardRepository extends JpaRepository<Board, Long> {
