+++ src/main/java/com/fcc/PureSync/context/board/repository/LikesRepository.java
@@ -1,11 +1,9 @@
-package com.fcc.PureSync.repository;
+package com.fcc.PureSync.context.board.repository;
 
-import com.fcc.PureSync.entity.Board;
-import com.fcc.PureSync.entity.Likes;
+import com.fcc.PureSync.context.board.entity.Board;
+import com.fcc.PureSync.context.board.entity.Likes;
 import com.fcc.PureSync.entity.Member;
-import org.apache.ibatis.annotations.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 public interface LikesRepository extends JpaRepository<Likes, Long> {
     Likes findByMemberAndBoard(Member member, Board board);
