+++ src/main/java/com/fcc/PureSync/service/MypageService.java
@@ -4,11 +4,12 @@ import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.fcc.PureSync.context.board.dto.BoardDto;
 import com.fcc.PureSync.dto.*;
 import com.fcc.PureSync.entity.*;
 import com.fcc.PureSync.exception.CustomException;
 import com.fcc.PureSync.exception.CustomExceptionCode;
-import com.fcc.PureSync.repository.BoardRepository;
+import com.fcc.PureSync.context.board.repository.BoardRepository;
 import com.fcc.PureSync.repository.BodyRepository;
 import com.fcc.PureSync.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
